Implement a transform.
Can be used to convert entries on push so that their true heft can be
determined. Realizign that pipelining should not be excessive. We
shouldn't pipe from one queue to the next with different strands
transforming as if it where UNIX pipes.

Also, probably want to add heft and back-pressure to Turnstile.